Fix/#135 diary기능 구현최종

+++ BE/src/main/java/com/oss/maeumnaru/diary/dto/DiaryAnalysisRequestDto.java
@@ -12,7 +12,7 @@ import java.time.LocalTime;
 @Builder
 public class DiaryAnalysisRequestDto {
 
-    private long emotionRate;
+    private float emotionRate;
 
     private long mealCount;
